requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==2.2.1
idna==2.6
oauthlib==2.0.6
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.3
requests==2.20.0
requests-oauthlib==0.8.0
six==1.11.0
social-auth-core==1.5.0
djangorestframework ==3.9.1
gunicorn
django-heroku
markdown      
django-filter
python-dotenv
django-rest-framework-social-oauth2
django-cors-headers
